requirements.txt
torch>=1.9.0
torchvision>=0.10.0
torchaudio>=0.9.0
numpy>=1.21.0
pandas>=1.3.0
matplotlib>=3.4.0
pyarrow>=5.0.0
ipykernel>=6.0.0
black[jupyter]>=21.5b2
scikit-learn>=0.24.2
kagglehub>=0.1.0
streamlit>=1.24.0
Pillow>=9.0.0